Flatten module structure
