+++ src/lib.rs
@@ -119,7 +119,7 @@ pub mod client;
 pub mod server;
 pub mod slave;
 
-pub use crate::service::service::{NewService, Service};
+pub use crate::service::{NewService, Service};
 
 mod codec;
 mod frame;
